% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbarplot.R
\name{fbarplot}
\alias{fbarplot}
\title{Make a bar plot for a factor}
\usage{
fbarplot(x, xlabel = "Levels", ylabel = "Count")
}
\arguments{
\item{x}{factor}

\item{xlabel}{character}

\item{ylabel}{character}
}
\value{
ggplot
}
\description{
Create a bar plot for a factor with user-defined axis names.
}
\examples{
x <- factor(sample(letters[1:5], size = 100, replace = TRUE))
fbarplot(x, xlabel="Letters")
}
